requirements.txt
torch>=2.0.0
transformers>=4.45.0
accelerate>=0.20.0
Pillow>=9.0.0
PyYAML>=6.0
